pyproject.toml
[tool.poetry]
name = "sknlp-serving"
version = "0.1.0"
description = ""
authors = ["nanaya <nanaya100@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
grpcio = "^1.42.0"
jieba = "^0.42.1"
scipy = "^1.7.2"
starlette = "^0.17.1"
uvicorn = "^0.15.0"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.42.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
